requirements.txt
httpx
uvicorn
pydantic
python-dotenv
web3
xrpl-py
fastapi
fastmcp
supabase
langchain
langchain-core
langchain-openai
langgraph
langgraph-checkpoint-postgres
psycopg[binary]
graphviz
langchain-mcp-adapters